requirements.txt
pyttsx3
streamlit 
python-dotenv
requests
openai
gtts
